Remove BaseSavableScene in favor of BasicScene
The BasicScene provides the SavableScene functionality without actually
implementing the SavableScene interface. Subclasses of BasicScene can opt
in to the state saving behavior by letting the subclass explicitly implement
SavableScene.
