+++ samples/notes-app/src/main/java/com/nhaarman/acorn/notesapp/presentation/createitem/CreateItemScene.kt
@@ -20,6 +20,7 @@ import com.nhaarman.acorn.notesapp.mainThread
 import com.nhaarman.acorn.notesapp.note.NoteItem
 import com.nhaarman.acorn.notesapp.note.NoteItemsRepository
 import com.nhaarman.acorn.presentation.RxScene
+import com.nhaarman.acorn.presentation.SavableScene
 import com.nhaarman.acorn.presentation.SceneKey.Companion.defaultKey
 import com.nhaarman.acorn.state.SceneState
 import io.reactivex.Observable
@@ -31,7 +32,8 @@ class CreateItemScene(
     private val noteItemsRepository: NoteItemsRepository,
     private val listener: Events,
     savedState: SceneState? = null
-) : RxScene<CreateItemContainer>(savedState) {
+) : RxScene<CreateItemContainer>(savedState),
+    SavableScene {
 
     private val textObservable by lazy {
         Observable.just(initialText ?: "")
