+++ samples/notes-app/src/main/java/com/nhaarman/acorn/notesapp/presentation/itemlist/ItemListScene.kt
@@ -20,6 +20,7 @@ import com.nhaarman.acorn.notesapp.mainThread
 import com.nhaarman.acorn.notesapp.note.NoteItem
 import com.nhaarman.acorn.notesapp.note.NoteItemsRepository
 import com.nhaarman.acorn.presentation.RxScene
+import com.nhaarman.acorn.presentation.SavableScene
 import com.nhaarman.acorn.presentation.SceneKey.Companion.defaultKey
 import com.nhaarman.acorn.state.SceneState
 import io.reactivex.rxkotlin.plusAssign
@@ -28,7 +29,8 @@ class ItemListScene(
     private val noteItemsRepository: NoteItemsRepository,
     private val listener: Events,
     savedState: SceneState?
-) : RxScene<ItemListContainer>(savedState) {
+) : RxScene<ItemListContainer>(savedState),
+    SavableScene {
 
     private val createClicks = view.whenAvailable { it.createClicks }
     private val itemClicks = view.whenAvailable { it.itemClicks }
